requirements.txt
datasets
evaluate
joblib
lightning
matplotlib
numpy
pandas
PyYAML
scipy
tabulate
tiktoken
torch
tqdm
transformers
wandb
gradio